+++ test/unit/nodejs/restWorkerTests.js
@@ -379,7 +379,7 @@ describe('restWorker', () => {
                 };
             });
 
-            it('should remove the revokeFrom property from the license after revoking', () => new Promise((resolve, reject) => {
+            it('should remove the revoke properties from the license after revoking', () => new Promise((resolve, reject) => {
                 const success = () => {};
                 const error = () => {
                     reject(new Error('should have called success'));
@@ -397,7 +397,8 @@ describe('restWorker', () => {
                                     Common: {
                                         myLicense: {
                                             class: 'License',
-                                            revokeFrom: 'foo'
+                                            revokeFrom: 'foo',
+                                            revokeCurrent: true
                                         }
                                     }
                                 }
@@ -410,7 +411,13 @@ describe('restWorker', () => {
                     try {
                         assert.strictEqual(
                             restWorker.state.doState.tasks[1234].internalDeclaration.Common.myLicense.revokeFrom,
-                            undefined
+                            undefined,
+                            'revokeFrom should be undefined'
+                        );
+                        assert.strictEqual(
+                            restWorker.state.doState.tasks[1234].internalDeclaration.Common.myLicense.revokeCurrent,
+                            undefined,
+                            'revokeCurrent should be undefined'
                         );
                         resolve();
                     } catch (err) {